# openapi.json
{
  "components": {
    "schemas": {
...
      "DatapointCreate": {
        "properties": {
          "capture_label": {
            "default": "",
            "title": "Capture Label",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/DatapointMetadata"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "title": "Points",
            "type": "array"
          },
          "protocol_version": {
            "default": "v1",
            "title": "Protocol Version",
            "type": "string"
          }
        },
        "required": [
          "points"
        ],
        "title": "DatapointCreate",
        "type": "object"
      },
      "DatapointListItem": {
        "properties": {
          "capture_label": {
            "title": "Capture Label",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "features": {
            "additionalProperties": true,
            "title": "Features",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/DatapointMetadata"
          },
          "protocol_version": {
            "title": "Protocol Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "capture_label",
          "protocol_version",
          "metadata",
          "features"
        ],
        "title": "DatapointListItem",
        "type": "object"
      },
      "DatapointMetadata": {
        "additionalProperties": true,
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "canvas": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CanvasInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "direction": {
            "anyOf": [
              {
                "enum": [
                  "clockwise",
                  "counterclockwise",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "hand": {
            "anyOf": [
              {
                "enum": [
                  "left",
                  "right",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hand"
          },
          "input_type": {
            "anyOf": [
              {
                "enum": [
                  "touch",
                  "pen",
                  "mouse",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Type"
          },
          "labels": {
            "additionalProperties": true,
            "title": "Labels",
            "type": "object"
          },
          "sampling": {
            "anyOf": [
              {
                "enum": [
                  "raw",
                  "time",
                  "distance",
                  "hybrid"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sampling"
          },
          "tool": {
            "anyOf": [
              {
                "enum": [
                  "index",
                  "thumb",
                  "stylus",
                  "mouse",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          }
        },
        "title": "DatapointMetadata",
        "type": "object"
      },
      "DatapointOut": {
        "properties": {
          "capture_label": {
            "title": "Capture Label",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "features": {
            "additionalProperties": true,
            "title": "Features",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/DatapointMetadata"
          },
          "protocol_version": {
            "title": "Protocol Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "capture_label",
          "protocol_version",
          "metadata",
          "features"
        ],
        "title": "DatapointOut",
        "type": "object"
      },
      "DatapointRawOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "raw": {
            "additionalProperties": true,
            "title": "Raw",
            "type": "object"
          }
        },
        "required": [
          "id",
          "raw"
        ],
        "title": "DatapointRawOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Point": {
        "properties": {
          "azimuth": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Azimuth"
          },
          "pressure": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressure"
          },
          "tMs": {
            "minimum": 0.0,
            "title": "Tms",
            "type": "integer"
          },
          "tiltX": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tiltx"
          },
          "tiltY": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tilty"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "tMs"
        ],
        "title": "Point",
        "type": "object"
      },
...
      "RewardEventOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "datapoint_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datapoint Id"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "points": {
            "title": "Points",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "points",
          "reason"
        ],
        "title": "RewardEventOut",
        "type": "object"
      },
      "RewardSummaryOut": {
        "properties": {
          "level": {
            "title": "Level",
            "type": "integer"
          },
          "next_level_at": {
            "title": "Next Level At",
            "type": "integer"
          },
          "points_per_level": {
            "title": "Points Per Level",
            "type": "integer"
          },
          "total_points": {
            "title": "Total Points",
            "type": "integer"
          }
        },
        "required": [
          "total_points",
          "level",
          "points_per_level",
          "next_level_at"
        ],
        "title": "RewardSummaryOut",
        "type": "object"
      },
      "TagCreate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          }
        },
        "title": "TagCreate",
        "type": "object"
      },
      "TagOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "datapoint_id": {
            "format": "uuid",
            "title": "Datapoint Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          }
        },
        "required": [
          "id",
          "datapoint_id",
          "created_at",
          "tag",
          "note"
        ],
        "title": "TagOut",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "refreshUrl": "/api/v1/auth/refresh",
            "scopes": {
              "admin": "Admin access.",
              "annotations": "Read, write, edit, delete annotations.",
              "me": "Read information about the current user."
            },
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Kruzinator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datapoints": {
      "get": {
        "operationId": "list_datapoints_api_v1_datapoints_get",
        "parameters": [
          {
            "in": "query",
            "name": "protocol_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Protocol Version"
            }
          },
          {
            "in": "query",
            "name": "capture_label",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capture Label"
            }
          },
          {
            "in": "query",
            "name": "created_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "in": "query",
            "name": "created_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          },
          {
            "in": "query",
            "name": "label_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label Key"
            }
          },
          {
            "in": "query",
            "name": "label_value",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label Value"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatapointListItem"
                  },
                  "title": "Response List Datapoints Api V1 Datapoints Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "List Datapoints",
        "tags": [
          "datapoints"
        ]
      },
      "post": {
        "operationId": "create_datapoint_api_v1_datapoints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatapointCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatapointOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "Create Datapoint",
        "tags": [
          "datapoints"
        ]
      }
    },
    "/api/v1/datapoints/{datapoint_id}": {
      "delete": {
        "operationId": "delete_datapoint_api_v1_datapoints__datapoint_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "datapoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Datapoint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "Delete Datapoint",
        "tags": [
          "datapoints"
        ]
      },
      "get": {
        "operationId": "get_datapoint_api_v1_datapoints__datapoint_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "datapoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Datapoint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatapointOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "Get Datapoint",
        "tags": [
          "datapoints"
        ]
      }
    },
    "/api/v1/datapoints/{datapoint_id}/raw": {
      "get": {
        "operationId": "get_datapoint_raw_api_v1_datapoints__datapoint_id__raw_get",
        "parameters": [
          {
            "in": "path",
            "name": "datapoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Datapoint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatapointRawOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "Get Datapoint Raw",
        "tags": [
          "datapoints"
        ]
      }
    },
    "/api/v1/datapoints/{datapoint_id}/tags": {
      "post": {
        "operationId": "add_tag_api_v1_datapoints__datapoint_id__tags_post",
        "parameters": [
          {
            "in": "path",
            "name": "datapoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Datapoint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "Add Tag",
        "tags": [
          "datapoints"
        ]
      }
    },
...
    "/api/v1/rewards/events": {
      "get": {
        "operationId": "list_my_reward_events_api_v1_rewards_events_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RewardEventOut"
                  },
                  "title": "Response List My Reward Events Api V1 Rewards Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "List My Reward Events",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/v1/rewards/me": {
      "get": {
        "operationId": "get_my_rewards_api_v1_rewards_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "Get My Rewards",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/v1/users/me": {
      "delete": {
        "operationId": "delete_me_api_v1_users_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "Delete Me",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "get_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ],
        "summary": "Get Me",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}